requirements.txt
openai
python-dotenv
chromadb
sentence-transformers
